Added most of the geoJson for the borders and added the salespartners
Added all of the geojson for the salespartners, still need to add for the representatives.
Added all the sales partners to the partnerlist.js file. Need to add better descriptions. And there is currently not working webpages for Dsolution or JNK indonesia. The coordinates for them might be wrong as well.

Added all the logos for the sales partners